apiDump. Note: THIS IS A BINARY-COMPATIBLE CHANGE:
Unit-coercion in blocking bridges are not supported with IR compiler, so Java code compiled against the legacy bridges will fail with `NoSuchMethodError`.

However, we've been providing the new bridges that are still supported by v2.7+ since v2.2, about half a year ago. So the actual group of affected users will be small.

Still, we are going to review feedback on the 2.7-M1 release to see if legacy bridges are still important.